+++ python-sdk/src/astro/databases/base.py
@@ -3,13 +3,16 @@ from __future__ import annotations
 import logging
 import warnings
 from abc import ABC
-from typing import Any, Callable, Mapping
+from typing import TYPE_CHECKING, Any, Callable, Mapping
 
 import pandas as pd
 import sqlalchemy
 from airflow.hooks.dbapi import DbApiHook
 from pandas.io.sql import SQLDatabase
 from sqlalchemy import column, insert, select
+
+if TYPE_CHECKING:
+    from sqlalchemy.engine.cursor import CursorResult
 from sqlalchemy.sql import ClauseElement
 from sqlalchemy.sql.elements import ColumnClause
 from sqlalchemy.sql.schema import Table as SqlaTable
@@ -92,7 +95,7 @@ class BaseDatabase(ABC):
         sql: str | ClauseElement = "",
         parameters: dict | None = None,
         **kwargs,
-    ):
+    ) -> CursorResult:
         """
         Return the results to running a SQL statement.
 
@@ -114,8 +117,12 @@ class BaseDatabase(ABC):
             )
             sql = kwargs.get("sql_statement")  # type: ignore
 
+        # We need to autocommit=True to make sure the query runs. This is done exclusively for SnowflakeDatabase's
+        # truncate method to reflect changes.
         if isinstance(sql, str):
-            result = self.connection.execute(sqlalchemy.text(sql), parameters)
+            result = self.connection.execute(
+                sqlalchemy.text(sql).execution_options(autocommit=True), parameters
+            )
         else:
             result = self.connection.execute(sql, parameters)
         return result
@@ -234,7 +241,7 @@ class BaseDatabase(ABC):
         table: BaseTable,
         file: File | None = None,
         dataframe: pd.DataFrame | None = None,
-        columns_names_capitalization: ColumnCapitalization = "lower",  # skipcq
+        columns_names_capitalization: ColumnCapitalization = "original",  # skipcq
     ) -> None:
         """
         Create a SQL table, automatically inferring the schema using the given file.
@@ -555,7 +562,6 @@ class BaseDatabase(ABC):
     ) -> None:
         """
         Append the source table rows into a destination table.
-        The argument `if_conflicts` allows the user to define how to handle conflicts.
 
         :param source_table: Contains the rows to be appended to the target_table
         :param target_table: Contains the destination table in which the rows will be appended