+++ sql-cli/sql_cli/__main__.py
@@ -6,27 +6,15 @@ from dotenv import load_dotenv
 from rich import print as rprint
 
 import sql_cli
-from sql_cli.connections import convert_to_connection, validate_connections
 from sql_cli.constants import DEFAULT_AIRFLOW_HOME, DEFAULT_DAGS_FOLDER
-from sql_cli.dag_generator import generate_dag
-from sql_cli.project import Project
-from sql_cli.run_dag import run_dag
-from sql_cli.utils.airflow import (
-    get_dag,
-    retrieve_airflow_database_conn_from_config,
-    set_airflow_database_conn,
-)
 
 load_dotenv()
-app = typer.Typer(add_completion=False)
+app = typer.Typer(add_completion=False, context_settings={"help_option_names": ["-h", "--help"]})
 
 
-def set_logger_level(log_level: int) -> None:
-    for name in logging.root.manager.loggerDict:
-        logging.getLogger(name).setLevel(log_level)
-
-
-set_logger_level(logging.ERROR)
+airflow_logger = logging.getLogger("airflow")
+airflow_logger.setLevel(logging.CRITICAL)
+airflow_logger.propagate = False
 
 
 @app.command(help="Print the SQL CLI version.")
@@ -54,15 +42,14 @@ def generate(
         None, dir_okay=True, metavar="PATH", help="(Optional) Default: current directory.", show_default=False
     ),
 ) -> None:
+    from sql_cli import cli
+    from sql_cli.project import Project
+
     project_dir_absolute = project_dir.resolve() if project_dir else Path.cwd()
     project = Project(project_dir_absolute)
     project.load_config(env)
 
-    dag_file = generate_dag(
-        directory=project.directory / project.workflows_directory / workflow_name,
-        dags_directory=project.airflow_dags_folder,
-    )
-    rprint("The DAG file", dag_file.resolve(), "has been successfully generated. 🎉")
+    cli.generate_dag(project, env, workflow_name)
 
 
 @app.command(
@@ -83,10 +70,22 @@ def validate(
         help="(Optional) Identifier of the connection to be validated. By default checks all the env connections.",
     ),
 ) -> None:
+    from sql_cli.connections import validate_connections
+    from sql_cli.project import Project
+    from sql_cli.utils.airflow import retrieve_airflow_database_conn_from_config, set_airflow_database_conn
+
     project_dir_absolute = project_dir.resolve() if project_dir else Path.cwd()
     project = Project(project_dir_absolute)
+    project.load_config(environment=env)
 
-    validate_connections(project=project, environment=env, connection_id=connection)
+    # Since we are using the Airflow ORM to interact with connections, we need to tell Airflow to use our airflow.db
+    # The usual route is to set $AIRFLOW_HOME before Airflow is imported. However, in the context of the SQL CLI, we
+    # decide this during runtime, depending on the project path and SQL CLI configuration.
+    airflow_meta_conn = retrieve_airflow_database_conn_from_config(project.directory / project.airflow_home)
+    set_airflow_database_conn(airflow_meta_conn)
+
+    rprint(f"Validating connection(s) for environment '{env}'")
+    validate_connections(connections=project.connections, connection_id=connection)
 
 
 @app.command(
@@ -111,12 +110,18 @@ def run(
     ),
     verbose: bool = typer.Option(False, help="Whether to show airflow logs", show_default=True),
 ) -> None:
+    from sql_cli import cli
+    from sql_cli.project import Project
+    from sql_cli.utils.airflow import (
+        get_dag,
+        retrieve_airflow_database_conn_from_config,
+        set_airflow_database_conn,
+    )
+
     project_dir_absolute = project_dir.resolve() if project_dir else Path.cwd()
     project = Project(project_dir_absolute)
     project.update_config(environment=env)
     project.load_config(env)
-
-    connections = {c["conn_id"]: convert_to_connection(c) for c in project.connections}
 
     # Since we are using the Airflow ORM to interact with connections, we need to tell Airflow to use our airflow.db
     # The usual route is to set $AIRFLOW_HOME before Airflow is imported. However, in the context of the SQL CLI, we
@@ -124,12 +129,9 @@ def run(
     airflow_meta_conn = retrieve_airflow_database_conn_from_config(project.directory / project.airflow_home)
     set_airflow_database_conn(airflow_meta_conn)
 
-    dag_file = generate_dag(
-        directory=project.directory / project.workflows_directory / workflow_name,
-        dags_directory=project.airflow_dags_folder,
-    )
+    dag_file = cli.generate_dag(project, env, workflow_name)
     dag = get_dag(dag_id=workflow_name, subdir=dag_file.parent.as_posix(), include_examples=False)
-    run_dag(dag, run_conf=project.airflow_config, connections=connections, verbose=verbose)
+    cli.run_dag(project, env, dag, verbose)
 
 
 @app.command(
@@ -180,6 +182,8 @@ def init(
         show_default=False,
     ),
 ) -> None:
+    from sql_cli.project import Project
+
     project_dir_absolute = project_dir.resolve() if project_dir else Path.cwd()
     project = Project(project_dir_absolute, airflow_home, airflow_dags_folder)
     project.initialise()