+++ sql-cli/sql_cli/exceptions.py
@@ -4,3 +4,19 @@ class DagCycle(Exception):
 
 class InvalidProject(Exception):
     """An exception raised when Project is invalid."""
+
+
+class EmptyDag(Exception):
+    """An exception raised when there are no SQL files within the DAG."""
+
+
+class SqlFilesDirectoryNotFound(Exception):
+    """An exception raised when the sql files directory does not exist."""
+
+
+class ConnectionFailed(Exception):
+    """An exception raised when the sql file's connection cannot be established."""
+
+    def __init__(self, *args: object, conn_id: str) -> None:
+        self.conn_id = conn_id
+        super().__init__(*args)