+++ sql-cli/sql_cli/project.py
@@ -6,7 +6,9 @@ from configparser import ConfigParser
 from pathlib import Path
 from typing import Any
 
-from sql_cli.configuration import Config
+from airflow.models.connection import Connection
+
+from sql_cli.configuration import Config, convert_to_connection
 from sql_cli.constants import DEFAULT_AIRFLOW_HOME, DEFAULT_DAGS_FOLDER, DEFAULT_ENVIRONMENT
 from sql_cli.exceptions import InvalidProject
 from sql_cli.utils.airflow import disable_examples
@@ -36,7 +38,7 @@ class Project:
         self.directory = directory
         self._airflow_home = airflow_home
         self._airflow_dags_folder = airflow_dags_folder
-        self.connections: list[dict[str, Any]] = []
+        self.connections: list[Connection] = []
 
     @property
     def airflow_home(self) -> Path:
@@ -85,9 +87,9 @@ class Project:
         config = config.from_yaml_to_config()
 
         if self._airflow_home is not None:
-            config.write_value_to_yaml("airflow", "home", str(self._airflow_home))
+            config.write_value_to_yaml("airflow", "home", str(self._airflow_home.resolve()))
         if self._airflow_dags_folder is not None:
-            config.write_value_to_yaml("airflow", "dags_folder", str(self._airflow_dags_folder))
+            config.write_value_to_yaml("airflow", "dags_folder", str(self._airflow_dags_folder.resolve()))
 
         config.connections[0]["host"] = str(self.directory / config.connections[0]["host"])
         config.write_config_to_yaml()
@@ -157,7 +159,7 @@ class Project:
             raise InvalidProject("This is not a valid SQL project. Please, use `flow init`")
         config = Config(environment=environment, project_dir=self.directory).from_yaml_to_config()
         if config.airflow_home:
-            self._airflow_home = Path(config.airflow_home)
+            self._airflow_home = Path(config.airflow_home).resolve()
         if config.airflow_dags_folder:
-            self._airflow_dags_folder = Path(config.airflow_dags_folder)
-        self.connections = config.connections
+            self._airflow_dags_folder = Path(config.airflow_dags_folder).resolve()
+        self.connections = [convert_to_connection(c) for c in config.connections]