+++ sql-cli/tests/test___main__.py
@@ -1,10 +1,12 @@
 import pathlib
+from tempfile import gettempdir
 
 import pytest
 from typer.testing import CliRunner
 
 from sql_cli import __version__
 from sql_cli.__main__ import app
+from sql_cli.connections import CONNECTION_ID_OUTPUT_STRING_WIDTH
 from tests.utils import list_dir
 
 runner = CliRunner()
@@ -54,22 +56,78 @@ def test_generate(workflow_name, environment, initialised_project):
             initialised_project.directory.as_posix(),
         ],
     )
-    if result.exit_code != 0:
-        print(result.output)
-    assert result.exit_code == 0
+    assert result.exit_code == 0, result.output
     result_stdout = get_stdout(result)
-    assert result_stdout.startswith("The DAG file ")
-    assert result_stdout.endswith(f"{workflow_name}.py has been successfully generated. 🎉")
+    assert (
+        f"The DAG file {initialised_project.airflow_dags_folder}/{workflow_name}.py has been successfully generated. 🎉"
+        in result_stdout
+    )
 
 
-def test_validate_with_directory(tmp_path):
-    result = runner.invoke(app, ["init", tmp_path.as_posix()])
-    assert result.exit_code == 0
-    result = runner.invoke(app, ["validate", tmp_path.as_posix()])
+@pytest.mark.parametrize(
+    "workflow_name,message",
+    [
+        ("non_existing", "The workflow non_existing does not exist!"),
+        ("cycle", "The workflow cycle contains a cycle! A cycle between d and d has been detected!"),
+        ("empty", "The workflow empty does not have any SQL files!"),
+    ],
+    ids=[
+        "non_existing",
+        "cycle",
+        "empty",
+    ],
+)
+def test_generate_invalid(workflow_name, message, initialised_project_with_tests_workflows):
+    result = runner.invoke(
+        app,
+        [
+            "generate",
+            workflow_name,
+            "--project-dir",
+            initialised_project_with_tests_workflows.directory.as_posix(),
+        ],
+    )
+    assert result.exit_code == 1
+    result_stdout = get_stdout(result)
+    assert message in result_stdout
+
+
+@pytest.mark.parametrize(
+    "env,connection,status",
+    [
+        ("default", "sqlite_conn", "PASSED"),
+        ("test", "sqlite_conn_invalid", "FAILED"),
+    ],
+)
+def test_validate(env, connection, status, initialised_project_with_test_config):
+    result = runner.invoke(
+        app,
+        [
+            "validate",
+            initialised_project_with_test_config.directory.as_posix(),
+            "--env",
+            env,
+            "--connection",
+            connection,
+        ],
+    )
+    assert result.exit_code == 0, result.exception
+    output = get_stdout(result)
+    assert f"Validating connection(s) for environment '{env}'" in output
+    assert f"Validating connection {connection:{CONNECTION_ID_OUTPUT_STRING_WIDTH}} {status}" in output
+
+
+def test_validate_all(initialised_project_with_test_config):
+    result = runner.invoke(
+        app,
+        [
+            "validate",
+            initialised_project_with_test_config.directory.as_posix(),
+        ],
+    )
     assert result.exit_code == 0
     output = get_stdout(result)
-    assert "Validating connection(s) for environment 'default'" in output
-    assert "Validating connection sqlite_conn               PASSED" in output
+    assert output.startswith("Validating connection(s)")
 
 
 @pytest.mark.parametrize(
@@ -91,11 +149,43 @@ def test_run(workflow_name, environment, initialised_project):
             initialised_project.directory.as_posix(),
         ],
     )
-    if result.exit_code != 0:
-        print(result.output)
-    assert result.exit_code == 0
+    assert result.exit_code == 0, result.output
     result_stdout = get_stdout(result)
-    assert f"Completed running the workflow {workflow_name}: [SUCCESS]" in result_stdout
+    assert f"Completed running the workflow {workflow_name}. 🚀" in result_stdout
+
+
+@pytest.mark.parametrize(
+    "workflow_name,message",
+    [
+        ("non_existing", "The workflow non_existing does not exist!"),
+        ("cycle", "The workflow cycle contains a cycle! A cycle between d and d has been detected!"),
+        ("empty", "The workflow empty does not have any SQL files!"),
+        ("undefined_variable", "'foo' is undefined"),
+        ("missing_table_or_conn_id", "You need to provide a table or a connection id"),
+        ("example_templating", "no such table: orders using connection sqlite_conn"),
+    ],
+    ids=[
+        "non_existing",
+        "cycle",
+        "empty",
+        "undefined_variable",
+        "missing_table_or_conn_id",
+        "example_templating",
+    ],
+)
+def test_run_invalid(workflow_name, message, initialised_project_with_tests_workflows):
+    result = runner.invoke(
+        app,
+        [
+            "run",
+            workflow_name,
+            "--project-dir",
+            initialised_project_with_tests_workflows.directory.as_posix(),
+        ],
+    )
+    assert result.exit_code == 1
+    result_stdout = get_stdout(result)
+    assert message in result_stdout
 
 
 def test_init_with_directory(tmp_path):
@@ -107,8 +197,9 @@ def test_init_with_directory(tmp_path):
 
 
 def test_init_with_custom_airflow_config(tmp_path):
+    tmp_dir = gettempdir()
     result = runner.invoke(
-        app, ["init", tmp_path.as_posix(), "--airflow-home", "/tmp", "--airflow-dags-folder", "/tmp"]
+        app, ["init", tmp_path.as_posix(), "--airflow-home", tmp_dir, "--airflow-dags-folder", tmp_dir]
     )
     assert result.exit_code == 0
     expected_msg = f"Initialized an Astro SQL project at {tmp_path.as_posix()}"